refactor: Rename batch interfaces to ItemReader/Processor/Writer
