+++ example/weather/resources/job.yaml
@@ -10,23 +10,23 @@ flow:
     fetchWeatherDataStep:
       id: fetchWeatherDataStep
       reader:
-        ref: weatherReader # JobFactory で登録されているコンポーネント名
+        ref: weatherItemReader # 新しいコンポーネント名に更新
       processor:
-        ref: weatherProcessor # JobFactory で登録されているコンポーネント名
+        ref: weatherItemProcessor # 新しいコンポーネント名に更新
       writer:
-        ref: weatherWriter # JobFactory で登録されているコンポーネント名
+        ref: weatherItemWriter # 新しいコンポーネント名に更新
       chunk:
         item-count: 50 # config.Batch.ChunkSize を利用
-        commit-interval: 50 # ★ 追加: コミット間隔を設定
+        commit-interval: 1 # ★ 追加: コミット間隔を設定
       listeners: # ステップレベルリスナー
         - ref: loggingStepListener
         - ref: retryStepListener
-      item-read-listeners: # アイテム読み込みリスナー
-        # - ref: myItemReadListener # 必要に応じて追加
-      item-process-listeners: # アイテム処理リスナー
-        # - ref: myItemProcessListener # 必要に応じて追加
-      item-write-listeners: # アイテム書き込みリスナー
-        # - ref: myItemWriteListener # 必要に応じて追加
+      # item-read-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: loggingItemReadListener
+      # item-process-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: loggingItemProcessListener
+      # item-write-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: loggingItemWriteListener
       skip-listeners: # アイテムスキップリスナー
         - ref: loggingSkipListener
       retry-item-listeners: # アイテムリトライリスナー
@@ -39,23 +39,23 @@ flow:
     processDummyDataStep:
       id: processDummyDataStep
       reader:
-        ref: dummyReader # JobFactory で登録されているダミーコンポーネント名
+        ref: dummyItemReader # 新しいコンポーネント名に更新
       processor:
-        ref: dummyProcessor # JobFactory で登録されているダミーコンポーネント名
+        ref: dummyItemProcessor # 新しいコンポーネント名に更新
       writer:
-        ref: dummyWriter # JobFactory で登録されているダミーコンポーネント名
+        ref: dummyItemWriter # 新しいコンポーネント名に更新
       chunk:
         item-count: 10 # このステップのチャンクサイズ
-        commit-interval: 10 # ★ 追加: コミット間隔を設定
+        commit-interval: 1 # ★ 追加: コミット間隔を設定
       listeners: # ステップレベルリスナー
         - ref: loggingStepListener
         - ref: retryStepListener
-      item-read-listeners: # アイテム読み込みリスナー
-        # - ref: myItemReadListener # 必要に応じて追加
-      item-process-listeners: # アイテム処理リスナー
-        # - ref: myItemProcessListener # 必要に応じて追加
-      item-write-listeners: # アイテム書き込みリスナー
-        # - ref: myItemWriteListener # 必要に応じて追加
+      # item-read-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: myItemReadListener
+      # item-process-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: myItemProcessListener
+      # item-write-listeners: # これらのリスナーはまだ登録されていないため、コメントアウトまたは削除を検討
+      #   - ref: myItemWriteListener
       skip-listeners: # アイテムスキップリスナー
         - ref: loggingSkipListener
       retry-item-listeners: # アイテムリトライリスナー
