+++ example/weather/step/processor/dummy_processor.go
@@ -1,22 +1,21 @@
-package processor
+package weatherprocessor // パッケージ名を 'weatherprocessor' に変更
 
 import (
 	"context"
+	itemprocessor "sample/pkg/batch/step/processor" // Renamed import
+	logger "sample/pkg/batch/util/logger"
 	"fmt" // Add fmt for Sprintf
 	"time"
-
-	processor "sample/pkg/batch/step/processor" // pkg/batch/step/processor を参照
-	logger "sample/pkg/batch/util/logger"
 )
 
 // DummyProcessor は入力アイテムをそのまま返すダミーの Processor です。
-// Processor[any, any] インターフェースを実装します。
+// ItemProcessor[any, any] インターフェースを実装します。
 type DummyProcessor struct{}
 
 // NewDummyProcessor は新しい DummyProcessor のインスタンスを作成します。
 func NewDummyProcessor() *DummyProcessor { return &DummyProcessor{} }
 
-// Process は Processor インターフェースの実装です。
+// Process は ItemProcessor インターフェースの実装です。
 // 入力として受け取ったアイテムをそのまま返すか、汎用的なダミーデータを返します。
 func (p *DummyProcessor) Process(ctx context.Context, item any) (any, error) { // I は any, O は any
 	// Context の完了をチェック
@@ -33,5 +32,5 @@ func (p *DummyProcessor) Process(ctx context.Context, item any) (any, error) { /
 	return any(fmt.Sprintf("Processed dummy item: %v at %s", item, time.Now().Format(time.RFC3339))), nil
 }
 
-// DummyProcessor が Processor[any, any] インターフェースを満たすことを確認
-var _ processor.Processor[any, any] = (*DummyProcessor)(nil)
+// DummyProcessor が ItemProcessor[any, any] インターフェースを満たすことを確認
+var _ itemprocessor.ItemProcessor[any, any] = (*DummyProcessor)(nil) // ここは itemprocessor を参照
