+++ pkg/batch/job/factory/jobfactory.go
@@ -10,9 +10,9 @@ import (
 	jsl "sample/pkg/batch/job/jsl"
 	jobListener "sample/pkg/batch/job/listener"
 	stepListener "sample/pkg/batch/step/listener" // Keep stepListener for step-level listeners
-	repository "sample/pkg/batch/repository"
 	logger "sample/pkg/batch/util/logger"
 	exception "sample/pkg/batch/util/exception"
+	repository "sample/pkg/batch/repository"
 )
 
 // ComponentBuilder は、特定のコンポーネント（Reader, Processor, Writer）を生成するための関数型です。
